requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
twilio
google-cloud-storage
python-multipart